### 2016-05-07 #42 @leeper
tabulizer: Bindings for Tabula PDF Table Extractor Library
* 1. What does this package do? This page wraps the [Tabula](tabula.technology) Java library, which can (very accurately) extract tables from PDF documents. It also implements some lower level utilities for working with PDF documents (metadata and text extraction, image conversion, split/merge). It should be useful for extracting scientific data, especially tabular data, from PDFs, such as from scientific articles or agency reports.
* 2. Paste the full DESCRIPTION file inside a code block below.

```
Package: tabulizer
Title: Bindings for Tabula PDF Table Extractor Library
Version: 0.1.11
Authors@R: c(person("Thomas J.", "Leeper", role = c("aut", "cre"),
                    email = "thosjleeper@gmail.com"))
Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
Description: Bindings for the Tabula <http://tabula.technology/> java library, which can extract tables from PDF documents.
    graphics,
    grDevices,
    tools,
    tabulizerjars,
    rJava,
    png
    knitr

```

* 3. URL for the package (the development repository, not a stylized html page)  https://github.com/leeper/tabulizer
* 4. What data source(s) does it work with (if applicable)?  n/a
* 5. Who is the target audience?  Data scientists stuck with other peoples' data
* 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?  The closest thing will be [pdftools](https://cran.rstudio.com/web/packages/pdftools/), which is a libpoppler wrapper. tabulizer has some overlap but the core functionality - table extraction - is not supported by pdftools.
* 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
    * [x] This package does not violate the Terms of Service of any service it interacts with.
    * [x] The repository has continuous integration with Travis CI and/or another service
    * [x] The package contains a vignette
    * [x] The package contains a reasonably complete README with `devtools` install instructions
    * [x] The package contains unit tests
    * [x] The package only exports functions to the NAMESPACE that are intended for end users
* 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
    * [] Are there any package dependencies not on CRAN?  The package has some files in a dependent package, which can be put on CRAN once everything is ready to release.
    * [x] Does the package have a CRAN accepted license?
    * [x] Did `devtools::check()` produce any errors or warnings? If so paste them below.
* 9. Please add explanations below for any exceptions to the above:
* 10. If this is a resubmission following rejection, please explain the change in circumstances.
