### 2016-05-07 f426668
add tabulizerjars install instructions [skip ci]
