### 2016-05-10 595fc95
update install instructions [skip ci]
