### 2016-05-10 #42 @sckott
Reviewers: @lmullmen
Due Date: 2016-06-01

Seeking another reviewer with Java experience ...
