### 2016-05-16 f7c5114
allow URLs in extract_areas()
