### 2016-05-24 #42 @davidgohel
This package is a very useful new package. It is also well written ; code is simple and clear, standards are respected. It was easy and fun to dive into the code. (Also I realized I could simplify some of my old
codes while reading yours!). :clap:

This work is a very good reference for package developpement with `rJava`.

I have tested it on:

-   mac OS (R version 3.2.4)
-   Windows (R version 3.2.3 32 bits) - locale is french Cp1252

Everything worked as expected. Installation is simple and the vignette is clear. I enjoy the `data.frame` method, it worked well on relevant documents.

There is nothing to complain about in the code. However, I made efforts to find things to say ;)

### Headless Mode

I found out one issue that can be easily solved. **Headless Mode**: it should be set to true by default (this can be an issue with RStudio server for example). If not, pdfbox might not work on some servers. See this [article](http://www.oracle.com/technetwork/articles/javase/headless-136834.html).

In `.onLoad` function, this can easily setup with

    J('java.lang.System')$setProperty('java.awt.headless', 'true')

### Encoding issue with Windows

I met one annoying issue. I made the *french* test, trying to use `tabulizer` with a document full of terrible accents. Everything went fine on mac, but I met encoding issues with Windows.

Note : I tried to set encoding parameters but it did not solve the issue.

    J('java.lang.System')$setProperty('file.encoding', 'UTF-8')

It seems the problem come from pdfbox and the way encodings are managed (not sure about that).

To demo the issue:

1.  Save manually the document from `http://www.sfds.asso.fr/ressource.php?fct=ddoc&i=2090` to `ZOOM_METIERS_MATHS_WEB_144dpi3.pdf` in your working directory.

2.  Run

    f <- "ZOOM_METIERS_MATHS_WEB_144dpi3.pdf"
    extract_text(f, pages = 2)

Below a partial copy of the result :

    [1] "ALAURÃ‰AT\r\nANNÃ‰E\r\n1\r\nANNÃ‰E\r\n2\r\nANNÃ‰E\r\n6\r\n3e annÃ©e\r\nANNÃ‰E\r\n8\r\nANNÃ‰E\r\n5\r\nÃ‰COLE\r\nD'INGÃ‰NIEURS\r\n2e annÃ©e\r\nANNÃ‰E\r\n7\r\nÃ‰COLE\r\nD'INGÃ‰NIEURS\r\nANNÃ‰E\r\n3\r\nANNÃ‰E\r\n4\r\nLICENCE PRO\r\nÃ‰COLE\r\nD'INGÃ‰NIEURS\r\nClasse prÃ©pa2e annÃ©e\r\nClasse prÃ©pa1re annÃ©e\r\nCycle \r\nintÃ©grÃ©\r\n1re annÃ©e\r\nCycle \r\nintÃ©grÃ©\r\n2e annÃ©e\r\nL3\r\nL2\r\nL1\r\n1re annÃ©e\r\nM2\r\nM1\r\nLICENCE\r\nMASTER\r\nDOCTORAT\r\nDIPLÃ”ME\r\nD'ING

Below a partial copy of the result on my mac :

    [1] "ALAURÉAT\nANNÉE\n1\nANNÉE\n2\nANNÉE\n6\n3e année\nANNÉE\n8\nANNÉE\n5\nÉCOLE\nD'INGÉNIEURS\n2e année\nANNÉE\n7\nÉCOLE\nD'INGÉNIEURS\nANNÉE\n3\nANNÉE\n4\nLICENCE PRO\nÉCOLE\nD'INGÉNIEURS\nClasse prépa2e année\nClasse prépa1re année\nCycle \nintégré\n1re année\nCycle \nintégré\n2e année\nL3\nL2\nL1\n1re année\nM2\nM1\nLICENCE\nMASTER\nDOCTORAT\nDIPLÔME\nD'ING

### Minor issues / suggestions

#### DESCRIPTION file

I think R-Core team could ask you to add authors of the java packages as contributor and copyright holder (that was my case for package `ReporteRsjars`) in package `tabulizerjars`. To prevent this issue when submitting to CRAN, I would add Manuel Aristarán in field `Authors@R` of the DESCRIPTION file as:

    person("Manuel", "Aristarán", role = c("ctb", "cph"), comment = "tabula java library")

#### Verbose logging

As a user, I would prefer not having the logger messages. If it can help, it can be switch off :

**Solution 1** : Desable ALL loggers

    J("java.util.logging.LogManager")$getLogManager()$reset()

**Solution 2** : Desable only pdfbox loggers

**BUT** it would require org.apache.log4j in the classpath.

    logManager <- rJava::J("java.util.logging.LogManager")$getLogManager()

    loggers <- c("org.apache.pdfbox.util.PDFStreamEngine",
      "org.apache.pdfbox.pdmodel.font.PDSimpleFont",
      "org.apache.pdfbox.pdmodel.font.PDFont",
      "org.apache.pdfbox.pdmodel.font.FontManager",
      "org.apache.pdfbox.pdfparser.PDFObjectStreamParser")

    for( logger_id in loggers ){
      logger <- logManager$getLogger(logger_id)
      logger$setLevel(J("org.apache.log4j.Level.OFF"))
    }
