### 2016-05-25 996fa78
some changes from rOpenSci review
