+++ NAMESPACE
@@ -10,6 +10,7 @@ export(locate_areas)
 export(make_thumbnails)
 export(merge_pdfs)
 export(split_pdf)
+export(stop_logging)
 import(tabulizerjars)
 importFrom(grDevices,dev.capabilities)
 importFrom(grDevices,dev.off)
