### 2016-05-25 leeper/tabulizer#10 @leeper
Handle non-latin encodings
This seems really challenging given the quirkiness of PDF format, but is the big issue to left to implement [from rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/42)
