### 2016-05-31 #42 @sckott
@lmullen @davidgohel
Due: 2016-06-01 - hey there, it's been 21 days, please get your review in soon, thanks :smiley_cat: (ropensci-bot)
