### 2016-05-31 57a5f33
support password-protected pdfs (closes #11)
