+++ R/get_page_dims.R
@@ -4,6 +4,7 @@
 #' @param file A character string specifying the path or URL to a PDF file.
 #' @param pages An optional integer vector specifying pages to extract from.
 #' @param doc Optionally, an rJava reference to a PDDocument Java object.
+#' @param password Optionally, a character string containing a user password to access a secured PDF.
 #' @details \code{get_n_pages} returns the page length of a PDF document. \code{get_page_dims} extracts the dimensions of specified pages in a PDF document. This can be useful for figuring out how to specify the \code{area} argument in \code{\link{extract_tables}}
 #' @return For \code{get_n_pages}, an integer. For \code{get_page_dims}, a list of two-element numeric vectors specifying the width and height of each page, respectively.
 #' @references \href{http://tabula.technology/}{Tabula}
@@ -20,8 +21,8 @@
 #' @importFrom rJava J new
 #' @seealso \code{\link{extract_tables}}, \code{\link{extract_text}}, \code{\link{make_thumbnails}}
 #' @export
-get_page_dims <- function(file, pages = NULL) {
-    pdfDocument <- load_doc(file)
+get_page_dims <- function(file, pages = NULL, password = NULL) {
+    pdfDocument <- load_doc(file, password = password)
     on.exit(pdfDocument$close())
     
     if (!is.null(pages)) {
@@ -39,9 +40,9 @@ get_page_dims <- function(file, pages = NULL) {
 
 #' @rdname get_page_dims
 #' @export
-get_n_pages <- function(doc, file) {
+get_n_pages <- function(doc, file, password = NULL) {
     if (!missing(file)) {
-        doc <- load_doc(file)
+        doc <- load_doc(file, password = password)
         on.exit(doc$close())
     }
     doc$getDocumentCatalog()$getAllPages()$size()
