+++ R/split_merge.R
@@ -1,8 +1,9 @@
 #' @rdname split_merge
 #' @title Split and merge PDFs
 #' @description Split PDF into separate pages or merge multiple PDFs into one.
-#' @param file For \code{merge_pdfs}, a character vector specifying the path to one or more local PDF files. For \code{split_pdf}, a character string specifying the path or URL to a PDF file.
+#' @param file For \code{merge_pdfs}, a character vector specifying the path to one or more \emph{local} PDF files. For \code{split_pdf}, a character string specifying the path or URL to a PDF file.
 #' @param outfile A character string specifying the path to the PDF file to create from the merged documents.
+#' @param password Optionally, a character string containing a user password to access a secured PDF. Currently, encrypted PDFs cannot be merged with \code{merge_pdfs}.
 #' @details \code{\link{split_pdf}} splits the file listed in \code{file} into separate one-page doucments. \code{\link{merge_pdfs}} creates a single PDF document from multiple separate PDF files.
 #' @return For \code{split_pdfs}, a character vector specifying the output file names, which are patterned after the value of \code{file}. For \code{merge_pdfs}, the value of \code{outfile}.
 #' @author Thomas J. Leeper <thosjleeper@gmail.com>
@@ -24,9 +25,9 @@
 #' @importFrom rJava J new
 #' @importFrom tools file_path_sans_ext
 #' @export
-split_pdf <- function(file) {
-    file <- localize_file(file)
-    pdfDocument <- load_doc(file)
+split_pdf <- function(file, password = NULL) {
+    file <- localize_file(file, copy = TRUE)
+    pdfDocument <- load_doc(file, password = password)
     on.exit(pdfDocument$close())
     splitter <- new(J("org.apache.pdfbox.util.Splitter"))
     splitArray <- splitter$split(pdfDocument)
@@ -46,8 +47,8 @@ split_pdf <- function(file) {
 #' @rdname split_merge
 #' @export
 merge_pdfs <- function(file, outfile) {
-    file <- localize_file(file)
     outfile <- path.expand(outfile)
+    file <- localize_file(file, copy = TRUE)
     merger <- new(J("org.apache.pdfbox.util.PDFMergerUtility"))
     merger$setDestinationFileName(outfile)
     lapply(file, merger$addSource)

[…]
