### 2016-05-31 f86cc5c
add missing password warning on encrypted pdfs
