### 2016-05-31 71c3b82
pass optional encodings to handle non-latin text (closes #10)
