+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library("testthat")
 library("tabulizer")
 
+stop_logging()
 test_check("tabulizer")
