+++ R/output.R
@@ -49,7 +49,7 @@ write_jsons <- function(tables, file, ...) {
     normalizePath(dirname(file))
 }
 
-list_matrices <- function(tables, ...) {
+list_matrices <- function(tables, encoding = NULL, ...) {
     out <- list()
     n <- 1L
     tablesIterator <- tables$iterator()
@@ -63,21 +63,24 @@ list_matrices <- function(tables, ...) {
                 out[[n]][i, j] <- tab$getCell(i-1L, j-1L)$getText()
             }
         }
+        if (!is.null(encoding)) {
+            Encoding(out[[n]]) <- encoding
+        }
         rm(tab)
         n <- n + 1L
     }
     out
 }
 
-list_characters <- function(tables, sep = "\t", ...) {
-    m <- list_matrices(tables, ...)
+list_characters <- function(tables, sep = "\t", encoding = NULL, ...) {
+    m <- list_matrices(tables, encoding = encoding, ...)
     lapply(m, function(x) {
         paste0(apply(x, 1, paste, collapse = sep), collapse = "\n")
     })
 }
 
-list_data_frames <- function(tables, sep = "\t", stringsAsFactors = FALSE, ...) {
-    char <- list_characters(tables = tables, sep = sep)
+list_data_frames <- function(tables, sep = "\t", stringsAsFactors = FALSE, encoding = NULL, ...) {
+    char <- list_characters(tables = tables, sep = sep, encoding = encoding)
     lapply(char, function(x) {
         o <- try(read.delim(text = x, stringsAsFactors = stringsAsFactors, ...))
         if (inherits(o, "try-error")) {
