+++ R/extract_tables.R
@@ -8,6 +8,7 @@
 #' @param spreadsheet A logical indicating whether to use Tabula's spreadsheet extraction algorithm. If \code{NULL} (the default), an automated assessment is made about whether it is appropriate.
 #' @param method A function to coerce the Java response object (a Java ArrayList of Tabula Tables) to some output format. The default method, \dQuote{matrices}, returns a list of character matrices. See Details for other options.
 #' @param password Optionally, a character string containing a user password to access a secured PDF.
+#' @param encoding Optionally, a character string specifying an encoding for the text, to be passed to the assignment method of \code{\link[base]{Encoding}}.
 #' @param \dots These are additional arguments passed to the internal functions dispatched by \code{method}.
 #' @details This function mimics the behavior of the Tabula command line utility. It returns a list of R character matrices containing tables extracted from a file by default. This response behavior can be changed by using the following options. \code{method = "character"} returns a list of single-element character vectors, where each vector is a tab-delimited, line-separate string of concatenated table cells. \code{method = "data.frame"} attempts to coerce the structure returned by \code{method = "character"} into a list of data.frames and returns character strings where this fails. \code{method = "csv"} writes the tables to comma-separated (CSV) files using Tabula's CSVWriter method in the same directory as the original PDF. \code{method = "tsv"} does the same but with tab-separated (TSV) files using Tabula's TSVWriter and \code{method = "json"} does the same using Tabula's JSONWriter method. The previous three methods all return the path to the directory containing the extract table files. \code{method = "asis"} returns the Java object reference, which can be useful for debugging or for writing a custom parser.
 #' \code{\link{extract_areas}} implements this functionality in an interactive mode allowing the user to specify extraction areas for each page.
@@ -49,6 +50,7 @@ function(file,
          spreadsheet = NULL,
          method = "matrix",
          password = NULL,
+         encoding = NULL,
          ...) {
 
     pdfDocument <- load_doc(file, password = password)
@@ -115,9 +117,9 @@ function(file,
            "csv" = write_csvs(tables, file = file, ...),
            "tsv" = write_tsvs(tables, file = file, ...),
            "json" = write_jsons(tables, file = file, ...),
-           "character" = list_characters(tables, ...),
-           "matrix" = list_matrices(tables, ...),
-           "data.frame" = list_data_frames(tables, ...),
+           "character" = list_characters(tables, encoding = encoding, ...),
+           "matrix" = list_matrices(tables, encoding = encoding, ...),
+           "data.frame" = list_data_frames(tables, encoding = encoding, ...),
            "asis" = tables,
            tables)
 }
