### 2016-05-31 #42 @leeper
Following up on the first review, I believe I have now addressed all issues:

1. rJava is configured to run in headless mode [on package load](https://github.com/leeper/tabulizerjars/blob/master/R/onLoad.R#L3).
2. I have added an `encoding` argument to [`extract_text()`](https://github.com/leeper/tabulizer/blob/master/R/extract_text.R#L6) and [`extract_tables()`](https://github.com/leeper/tabulizer/blob/master/R/extract_tables.R#L11) that allows the specification of text encodings. I'm not 100% confident in this, but it seems to work.
3. I've updated authorship in [DESCRIPTION of tabulizerjars](https://github.com/leeper/tabulizerjars/blob/master/DESCRIPTION#L8-L9) to reflect the java library's authorship. I've also added @davidgohel's [contributions as a reviewer](https://github.com/leeper/tabulizer/blob/master/DESCRIPTION#L8-L9).
4. I use `path.expand()` [for local file paths](https://github.com/leeper/tabulizer/blob/master/R/utils.R#L11).
5. I have added [a `stop_logging()` function](https://github.com/leeper/tabulizer/blob/master/R/logging.R) to turn off all Java logging. This seemed easier than adding a log4j dependency, but may not be ideal, so logging is left on by default.

Thank you again for the feedback!
