### 2016-06-01 #42 @lmullen
This is a very well done package. The code is a pleasure to read, and the interface is well thought out. The core functionality works as expected.

This package is so useful that I would really love it if this package worked on PDFs of 19th-century documents: alas, it's not to be. (Hardly the fault of the package, of course.)

@davidgohel has already offered a number of suggestions which have been dealt with, so I'll deal mostly with unexpected errors when using the interface.

## Tests

I've run R CMD check on this package with R 3.3.0 on Mac OS X 10.11.5 and R 3.3.0 on Ubuntu 16.04. (The rest of my review was on the Mac. I don't have a Windows machine to test that installation, but the instructions in the README seemed clear.)

- The Travis integration on the master branch is failing.

- I get two test failures when building and running R CMD check (on both Mac and Ubuntu):

```
  OK: 73 SKIPPED: 0 FAILED: 2
  1. Failure: Read French language PDF w/correct encoding (@test_non-latin.R#17)
  2. Failure: Read French language PDF w/correct encoding (@test_non-latin.R#18)
```

I gather that this is part of the issue which @davidgohel mentioned which is being worked out.

## Interface

For testing I used a PDF form the U.S. Census's [Statistical Abstract](http://www2.census.gov/library/publications/2010/compendia/statab/129ed/tables/educ.pdf)

Some minor questions:

- Is there are reason why `get_n_pages()` takes an option `doc =` parameter first instead of the `file = ` parameter, like most other functions in this package?

- Can `split_pdf()` take an argument like `outdir = ` to specify where the split files should go? Also, it would be nice if the numbers in the new file names were zero padded. More importantly, `split_pdf()` overwrites files in the directory without warning. The function needs some way of requiring the user's permission before overwriting files, especially since the function doesn't let the user specify the file name or path.

- When using `merge_pdfs()` I get a warning message:

```
Warning message:
In if (grepl("^http.*://", path)) { :
  the condition has length > 1 and only the first element will be used
```

This comes from `tabulizer:::localize_file()`. The `if()` condition there probably needs to be `if(any(CONDITION))`. I suppose a user could pass the function a mix of remote and local PDFs so that the `any()` function would be insufficient, but that seems far-fetched.

- I get an error with a traceback when using the option `spreadsheet = TRUE` to `extract_tables()`.

```
 Error in .jcall("RJavaTools", "Ljava/lang/Object;", "invokeMethod", cl,  :
  java.lang.IndexOutOfBoundsException: Index: 0, Size: 0
```

If this isn't a problem with the underlying Java code, then this error message should let the user know why the spreadsheet option shouldn't be used.

- Same comments for `make_thumbnails()` as for `split_pdf()` above.

After playing with it for a while, I was able to figure out the interface to `extract_areas()`. Would it be possible to make two improvements to that user interface? First, instead of clicking and the upper left and lower right of the area, can the user draw a bounding box? Second, can the bounding box be adjusted before the function returns?

## Miscellaneous

- The coding style and documentation are very well done.

- I found it a little odd that the README had the installation instructions near the bottom, since it goes against the convention. Can you move the installation instructions up (maybe leaving the Windows Java stuff at the bottom)? See the README section of the [packaging guide](https://github.com/ropensci/packaging_guide).

- The package has clear contributing guidelines. It would be good to also have the code of conduct as mentioned in the rOpenSci policies.

- The package has a good NEWS file. But it also needs to use tags for all the point releases. I don't see any point in going back to tag the previous versions, but it should be tagged starting with acceptance at rOpenSci.

- My two cents: I'd much prefer if the Java logging was off by default. I don't think it tells most users anything that they can actually use.

Nice work, and I look forward to using the package myself.
