### 2016-06-01 #42 @leeper
Thank you, @lmullen! I will work to address these points and post an update once I have finished.
