### 2016-06-01 c65c64e
temporarily hold off on non-latin tests
