### 2016-06-01 c4bae59
further updates from rOpenSci onboarding
