+++ R/onAttach.R
@@ -1,0 +1,3 @@
+.onAttach <- function(libname, pkgname) {
+    stop_logging()
+}
