+++ R/get_page_dims.R
@@ -3,7 +3,7 @@
 #' @description Get Page Length and Dimensions
 #' @param file A character string specifying the path or URL to a PDF file.
 #' @param pages An optional integer vector specifying pages to extract from.
-#' @param doc Optionally, an rJava reference to a PDDocument Java object.
+#' @param doc Optionally,, in lieu of \code{file}, an rJava reference to a PDDocument Java object.
 #' @param password Optionally, a character string containing a user password to access a secured PDF.
 #' @details \code{get_n_pages} returns the page length of a PDF document. \code{get_page_dims} extracts the dimensions of specified pages in a PDF document. This can be useful for figuring out how to specify the \code{area} argument in \code{\link{extract_tables}}
 #' @return For \code{get_n_pages}, an integer. For \code{get_page_dims}, a list of two-element numeric vectors specifying the width and height of each page, respectively.
@@ -21,17 +21,19 @@
 #' @importFrom rJava J new
 #' @seealso \code{\link{extract_tables}}, \code{\link{extract_text}}, \code{\link{make_thumbnails}}
 #' @export
-get_page_dims <- function(file, pages = NULL, password = NULL) {
-    pdfDocument <- load_doc(file, password = password)
-    on.exit(pdfDocument$close())
+get_page_dims <- function(file, doc, pages = NULL, password = NULL) {
+    if (!missing(file)) {
+        doc <- load_doc(file, password = password)
+        on.exit(doc$close())
+    }
     
     if (!is.null(pages)) {
         pages <- as.integer(pages)
     } else {
-        pages <- 1L:(get_n_pages(doc = pdfDocument))
+        pages <- 1L:(get_n_pages(doc = doc))
     }
     
-    allpages <- pdfDocument$getDocumentCatalog()$getAllPages()
+    allpages <- doc$getDocumentCatalog()$getAllPages()
     lapply(pages, function(x) {
         thispage <- allpages$get(x-1L)
         c(thispage$getMediaBox()$getWidth(), thispage$getMediaBox()$getHeight())
@@ -40,7 +42,7 @@ get_page_dims <- function(file, pages = NULL, password = NULL) {
 
 #' @rdname get_page_dims
 #' @export
-get_n_pages <- function(doc, file, password = NULL) {
+get_n_pages <- function(file, doc, password = NULL) {
     if (!missing(file)) {
         doc <- load_doc(file, password = password)
         on.exit(doc$close())
