+++ man/stop_logging.Rd
@@ -13,7 +13,7 @@ stop_logging()
 Toggle verbose rJava logging
 }
 \details{
-This function turns off the default, somewhat verbose rJava logging, most of which is uninformative.
+This function turns off the somewhat verbose rJava logging, most of which is uninformative. It is called automatically when tabulizer is attached via \code{library()}, \code{require}, etc. To keep logging on, load the package namespace using \code{requireNamespace("tabulizer")} and reference functions in using fully qualified references (e.g., \code{tabulizer::extract_tables()}.
 }
 \note{
 This resets a global Java setting and may affect logging of other rJava operations, requiring a restart of R.
