### 2016-06-01 d68c9d0
finish rOpenSci onboarding updates
