### 2016-06-01 leeper/tabulizer#10 @leeper
This works on Windows, but is screwy on Linux. Must investigate more.
