### 2016-06-01 #42 @leeper
I have attempted to address all of your comments and have made the following revisions:

1. I have [stopped the failing encoding tests](https://github.com/leeper/tabulizer/blob/master/tests/testthat/test_non-latin.R#L17-L18) (but have [reopened an issue](https://github.com/leeper/tabulizer/issues/10) to get this to work better cross-platform).
2. I have [reordered the `file` and `doc` arguments](https://github.com/leeper/tabulizer/blob/master/R/get_page_dims.R#L24) to `get_n_pages()` for consistency and [given `get_page_dims()` a `doc` argument](https://github.com/leeper/tabulizer/blob/master/R/get_page_dims.R#L45), as well. I figure these may be useful for debugging.
3. [`split_pdf()`](https://github.com/leeper/tabulizer/blob/master/R/split_merge.R#L5) and [`make_thumbnails()`](https://github.com/leeper/tabulizer/blob/master/R/make_thumbnails.R#L4) gain an `outdir` argument. In both cases, output files are numbered using zero padding, as requested.
4. I have fixed the `merge_pdfs()` warnings by [avoiding passing a vector of filenames through `localize_file()`](https://github.com/leeper/tabulizer/blob/master/R/split_merge.R#L59).
5. The error related to the `spreadsheet = TRUE` argument [has been corrected](https://github.com/leeper/tabulizer/blob/master/R/output.R#L57-L60) and was quite critical, actually, as it related to the attempt to coerce an empty Java array to an R matrix.
6. I have turned [Java logging off by default](https://github.com/leeper/tabulizer/blob/master/R/onAttach.R) and [added some notes](https://github.com/leeper/tabulizer/blob/master/R/logging.R#L3-L4) to the documentation to describe how to restore logging if that were, for some reason, preferred.
7. Regarding the behavior of `extract_areas()`, I agree this is less than ideal but the current implementation isn't actually as strict as the documentation implies (you can tag upper-left/lower-right or upper-right/lower-left). I have [an open issue](https://github.com/leeper/tabulizer/issues/8) on improving this, but I think there are some fundamental limitations in `locator()` and graphics devices more generally for creating an intuitive interface since it is not really possible to dynamically highlight/select an area of a plot (as far as I know). I am open to suggestions on this.

On smaller points:

1. Regarding installation instructions, I think we're getting to the point where GitHub-based installation is sufficiently common that it doesn't need to be at the top of the README (since that seems to be a legacy of earlier times). I would actually suggest the general style guide be changed, accordingly. But I don't feel strongly about this, and can update if preferred.
2. I read the code of conduct provision as only a recommendation, so I have not added this.
3. My release workflow is to only tag "minor" not "patch" versions as GitHub releases and generally to only release to CRAN when minor ticks up, but I can adapt this to a different style if required.

Thank you, again, @davidgohel and @lmullen, for your feedback! It is very much appreciated.
