### 2016-06-01 #42 @lmullen
This all sounds good to me, @leeper. My only suggestion is that tags are cheap so it doesn't hurt to add them.

@sckott and @karthik and @cboettig: I recommend accepting this package to rOpenSci.
