### 2016-06-01 #42 @sckott
a few things:

* before everything's on CRAN...seems like using `Remotes` in `DESCRIPTION` for `tabulizerjars` would make it so that would automatically install for users
* After installing ``tabulizerjars`- Installation failed still, but after reinstalling `rJava` from source from CRAN, then it worked
* Getting errors on `extract_areas` and `locate_areas`, same error in each case:

```r
#> Error in (function (title, width, height, pointsize, family, antialias,  :
#>   unable to create quartz() device target, given type may not be supported
#> In addition: Warning message:
#> In (function (title, width, height, pointsize, family, antialias,  :
#>   No displays are available
```

I guess lincoln and david didn't get this error though, so maybe i'm just an edge case
