### 2016-06-02 #42 @sckott
yeah, from CRAN - I imagine we'd want this to work regardless of where R was installed from?
