### 2016-06-02 849c5da
work on locate_areas updates (#8)
