### 2016-06-02 #42 @leeper
Thanks, @sckott. I'm working on these issues now.
