### 2016-06-03 1a58bb8
completely redo locate_areas() (#8)
