+++ R/utils.R
@@ -60,7 +60,11 @@ make_area <- function(area = NULL, pages = NULL, npages = NULL) {
             }
         }
         area <- lapply(area, function(x) {
-            new(J("technology.tabula.Rectangle"), .jfloat(x[1]), .jfloat(x[2]), .jfloat(x[4]-x[2]), .jfloat(x[3]-x[1]))
+            if (!is.null(x)) {
+                new(J("technology.tabula.Rectangle"), .jfloat(x[1]), .jfloat(x[2]), .jfloat(x[4]-x[2]), .jfloat(x[3]-x[1]))
+            } else {
+                NULL
+            }
         })
     }
     area
