### 2016-06-03 #42 @leeper
I've just pushed an update to the `extract_areas()`/`locate_areas()` functions, which I hope might work better on Mac (not tested) and that offers much more interactive functionality on Windows (tested) and probably Linux (not tested yet) with the caveat that RStudio's graphic device apparently doesn't play well with others. Would any of you be willing to give this a quick run and see how it works for you?
