### 2016-06-03 #42 @davidgohel
Hi

I've just tested on mac. There is a first issue in `extract_areas` (line 96) . It shoud be `grDevices::X11(type = "Xlib")` and not `grDevices::X11(type = "xlib")`.

But then I got a

```
Error in try_area(file = paths[i], dims = dims[[i]], area = areas[[i]],  :
  Graphics device does not support rasterImage() plotting
```

Here is what my dev.cap are:

````
> grDevices::dev.capabilities()
$semiTransparency
[1] FALSE

$transparentBackground
[1] "fully"

$rasterImage
[1] "non-missing"

$capture
[1] TRUE

$locator
[1] TRUE

$events
[1] "MouseDown" "MouseMove" "MouseUp"   "Keybd"
````

and my session info:

```
> sessionInfo()
R version 3.3.0 (2016-05-03)
Platform: x86_64-apple-darwin13.4.0 (64-bit)
Running under: OS X 10.11.5 (El Capitan)

locale:
[1] fr_FR.UTF-8/fr_FR.UTF-8/fr_FR.UTF-8/C/fr_FR.UTF-8/fr_FR.UTF-8

attached base packages:

other attached packages:
[1] tabulizer_0.1.18

[1] tabulizerjars_0.1.2 tools_3.3.0         rJava_0.9-8         png_0.1-7
```
