### 2016-06-03 #42 @davidgohel
And all is ok when changing line 98 to

```
if (grDevices::dev.capabilities()[["rasterImage"]] == "no") {
```
