### 2016-06-03 #42 @sckott
Same problem, getting an error with `try_area`
