### 2016-06-07 c03ed0a
add/prefer RStudio widget for locate_areas() (#8)
