+++ R/locate_area.R
@@ -91,118 +91,14 @@ extract_areas <- function(file, pages = NULL, guess = FALSE, ...) {
 }
 
 try_area <- function(file, dims, area = NULL, warn = FALSE) {
-    deviceUnits <- "nfc"
-    if (Sys.info()["sysname"] == "Darwin") {
-        grDevices::X11(type = "xlib")
-    }
-    if (grDevices::dev.capabilities()[["rasterImage"]] != "yes") {
-        stop("Graphics device does not support rasterImage() plotting")
-    }
-    thispng <- readPNG(file, native = TRUE)
-    drawPage <- function() {
-        graphics::plot(c(0, dims[1]), c(0, dims[2]), type = "n", xlab = "", ylab = "", asp = 1)
-        graphics::rasterImage(thispng, 0, 0, dims[1], dims[2])
-    }
-    drawRectangle <- function() {
-        if (!is.null(endx)) {
-            graphics::rect(startx, starty, endx, endy, col = grDevices::rgb(1,0,0,.2) )
-        }
-    }
-        
-    pre_par <- graphics::par(mar=c(0,0,0,0), xaxs = "i", yaxs = "i", bty = "n")
-    on.exit(graphics::par(pre_par), add = TRUE)
-    drawPage()
-    on.exit(grDevices::dev.off(), add = TRUE)
-    
-    if (!length(grDevices::dev.capabilities()[["events"]])) {
-        if (warn) {
-            message("Graphics device does not support event handling...\n",
-                    "Entering reduced functionality mode.\n",
-                    "Click upper-left and then lower-right corners of area.")
-        }
-        tmp <- locator(2)
-        graphics::rect(tmp$x[1], tmp$y[1], tmp$x[2], tmp$y[2], col = grDevices::rgb(1,0,0,.5))
-        Sys.sleep(2)
-        
-        # convert to: top,left,bottom,right
-        area <- c(dims[2] - max(tmp$y), min(tmp$x), dims[2] - min(tmp$y), max(tmp$x))
-        return(list(key = "right", area = area))
-    }
-    
-    clicked <- FALSE
-    lastkey <- NA_character_
-    if (!length(area)) {
-        startx <- NULL
-        starty <- NULL
-        endx <- NULL
-        endy <- NULL
+    requireNamespace("rstudioapi")
+    if (rstudioapi::isAvailable()) {
+        try_area_rstudio(file = file, dims = dims, area = area)
     } else {
-        showArea <- function() {
-            # convert from: top,left,bottom,right
-            startx <<- area[2]
-            starty <<- dims[2] - area[1]
-            endx <<- area[4]
-            endy <<- dims[2] - area[3]
-            drawRectangle()
-        }
-        showArea()
-    }
-    
-    devset <- function() {
-        if (grDevices::dev.cur() != eventEnv$which) grDevices::dev.set(eventEnv$which)
-    }
-    
-    mousedown <- function(buttons, x, y) {
-        devset()
-        if (clicked) {
-            endx <<- graphics::grconvertX(x, deviceUnits, "user")
-            endy <<- graphics::grconvertY(y, deviceUnits, "user")
-            clicked <<- FALSE
-            eventEnv$onMouseMove <- NULL
+        if (!length(grDevices::dev.capabilities()[["events"]])) {
+            try_area_reduced(file = file, dims = dims, area = area, warn = warn)
         } else {
-            startx <<- graphics::grconvertX(x, deviceUnits, "user")
-            starty <<- graphics::grconvertY(y, deviceUnits, "user")
-            clicked <<- TRUE
-            eventEnv$onMouseMove <- dragmousemove
-        }
-        NULL
-    }
-
-    dragmousemove <- function(buttons, x, y) {
-        devset()
-        if (clicked) {
[…]
