+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: tabulizer
 Type: Package
 Title: Bindings for Tabula PDF Table Extractor Library
-Version: 0.1.18
-Date: 2016-06-03
+Version: 0.1.19
+Date: 2016-06-04
 Authors@R: c(person("Thomas J.", "Leeper", role = c("aut", "cre"),
                     email = "thosjleeper@gmail.com"),
              person("David", "Gohel", role = "ctb",
@@ -22,6 +22,9 @@ Imports:
 Suggests:
     graphics,
     grDevices,
+    rstudioapi,
+    shiny,
+    miniUI,
     testthat,
     knitr
 RoxygenNote: 5.0.1
