### 2016-06-07 8732d9c
set plot device margins in RStudio (#8) [skip ci]
