+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: tabulizer
 Type: Package
 Title: Bindings for Tabula PDF Table Extractor Library
-Version: 0.1.19
-Date: 2016-06-04
+Version: 0.1.20
+Date: 2016-06-08
 Authors@R: c(person("Thomas J.", "Leeper", role = c("aut", "cre"),
                     email = "thosjleeper@gmail.com"),
              person("David", "Gohel", role = "ctb",
