### 2016-06-08 bcc4193
document changes
