+++ R/try_area_methods.R
@@ -70,7 +70,7 @@ try_area_reduced <- function(file, dims, area = NULL, warn = FALSE) {
                 "Entering reduced functionality mode.\n",
                 "Click upper-left and then lower-right corners of area.")
     }
-    if (grDevices::dev.capabilities()[["rasterImage"]] != "yes") {
+    if (grDevices::dev.capabilities()[["rasterImage"]] != "no") {
         stop("Graphics device does not support rasterImage() plotting")
     }
     thispng <- readPNG(file, native = TRUE)
@@ -97,7 +97,7 @@ try_area_full <- function(file, dims, area = NULL) {
     
     deviceUnits <- "nfc"
     if (Sys.info()["sysname"] == "Darwin") {
-        grDevices::X11(type = "xlib")
+        grDevices::X11(type = "Xlib")
     }
     if (grDevices::dev.capabilities()[["rasterImage"]] != "yes") {
         stop("Graphics device does not support rasterImage() plotting")
