### 2016-06-08 #42 @leeper
Thank you everyone for the help so far. I've made some pretty radical changes to the `extract_areas()` functionality to try to make sure this works across platforms. The biggest change is that when working in RStudio a shiny widget is provided to handle the extraction directly within RStudio. In other cases, it falls back on the graphics device-based functionality described previously (and hopefully fixed due to @davidgohel's testing) and, if graphics events are unavailable on the device, falls back on using `locator()`, and in the (hopefully rare case that even that is unavailable), fails entirely.
