### 2016-06-08 #42 @sckott
I still get the same quartz error as above - haven't been able to google any solutions for this, weird
