### 2016-06-08 #42 @leeper
Ah, that's frustrating. You're on OS X, right? What does `Sys.info()["sysname"]` return?
