### 2016-06-09 #42 @lmullen
👍 👍 👍 for the new Shiny interface.

Any particular reason why you can't use the Shiny interface outside of RStudio and just open the gadget in a browser instead of the Viewer?
