### 2016-06-09 #42 @sckott
> use the Shiny interface outside of RStudio and just open the gadget in a browser instead of the Viewer?

yeah, should be possible
