### 2016-06-19 259f5ed
allow locate_areas() widget choice
