### 2016-06-19 #42 @leeper
I've now made the shiny interface the default, and added an option `widget` to control which interface is used for `locate_areas()` (and thus to `extract_areas()`). Sorry the delay in getting to it.
