### 2016-06-21 #42 @sckott
weird, now I'm getting the java error `error: unable to load shared object '/Library/Frameworks/R.framework/Versions/3.3/Resources/library/rJava/libs/rJava.so':` - will try to track that down
