+++ .travis.yml
@@ -1,5 +1,19 @@
 language: r
 sudo: required
+matrix:
+  include:
+    - os: linux
+      dist: precise
+      sudo: false
+    - os: linux
+      dist: trusty
+      env: R_CODECOV=true
+      r_check_args: '--use-valgrind'
+    - os: osx
+      osx_image: xcode8
+    - os: osx
+      osx_image: beta-xcode6.1
+      disable_homebrew: true
 r_packages:
 - knitr
 - covr
