### 2016-08-16 c352734
some tweaks to os x builds
