+++ .travis.yml
@@ -11,7 +11,7 @@ matrix:
     - os: osx
       osx_image: xcode8
     - os: osx
-      osx_image: beta-xcode6.1
+      osx_image: beta-xcode6.2
 r_packages:
 - knitr
 - covr
