### 2016-08-16 2bdc12d
rm unnecessary env var [skip ci]
