### 2016-08-16 6e3000f
add appveyor shield [skip ci]
