+++ README.Rmd
@@ -87,6 +87,7 @@ Some other utility functions are also provided (and made possible by the [Java P
 
 [![CRAN](http://www.r-pkg.org/badges/version/tabulizer)](http://cran.r-project.org/package=tabulizer)
 [![Build Status](https://travis-ci.org/leeper/tabulizer.png?branch=master)](https://travis-ci.org/leeper/tabulizer)
+[![Build status](https://ci.appveyor.com/api/projects/status/wj5afxb1v42h8oui?svg=true)](https://ci.appveyor.com/project/leeper/tabulizer)
 [![codecov.io](http://codecov.io/github/leeper/tabulizer/coverage.svg?branch=master)](http://codecov.io/github/leeper/tabulizer?branch=master)
 
 tabulizer depends on [rJava](https://cran.r-project.org/package=rJava), which implies a system requirement for Java. This can be frustrating, especially on Windows. My preferred setup routine is to use [Chocolatey](https://chocolatey.org/) to obtain, configure, and update Java (see instructions below). You need do this before installing rJava or attempting to use tabulizer.
