### 2016-08-16 c04dd5e
Update appveyor.yml
