+++ appveyor.yml
@@ -9,7 +9,7 @@ environment:
   - WARNINGS_ARE_ERRORS: 1
   - USE_RTOOLS: true
 
- matrix:
+matrix:
   - R_VERSION: devel
     GCC_PATH: mingw_32
 
