### 2016-08-16 eba17a8
Update appveyor.yml
