### 2016-08-16 3e72e0d
install tabulizerjars from GitHub on appveyor
