+++ appveyor.yml
@@ -28,6 +28,7 @@ install:
   ps: Bootstrap
 
 build_script:
+  - travis-tool.sh install_github leeper/tabulizerjars
   - travis-tool.sh install_deps
 
 test_script:
