### 2016-08-16 27dbb2a
set JAVA_HOME envvar on appveyor
