### 2016-08-16 74a7f33
maybe set JAVA_HOME to blank?
