### 2016-08-17 #42 @sckott
approved!

ran `goopractice` on it, not mandatory, but things to consider

```
It is good practice to
  ✖ omit "Date" in DESCRIPTION. It is not required and it
    gets invalid quite often. A build date will be added to the package
    when you perform `R CMD build` on it.

    R/extract_tables.R:78:1
    R/extract_tables.R:79:1
    R/extract_tables.R:98:1
    R/locate_area.R:45:1
    R/locate_area.R:56:1
    ... and 52 more lines

    vapply() instead.

    R/split_merge.R:59:13
    R/utils.R:38:5

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/output.R:65:19
    R/output.R:66:23
```
