### 2017-12-23 #42 @lmullen
@ajduncanson This is not the right place for support. Try the tabulizer repo itself.

That said, the problem is probably not with tabulizer. The time zone warning sounds like the one with a recent version of R on MacOS, and you may have warnings set to errors. I suggest that you search for a solution to the time zone problem, check whether you have warnings set to errors, and if neither of those work, try an tabulizer repo.
